+++ src/main/kotlin/environment/libEx/populateSnapShot.kt
@@ -4,7 +4,7 @@ import data.registers.RegisterType
 import environment.ExecuteLib
 import registers
 
-fun ExecuteLib.populateSnapShot(snapShotRegisters: MutableMap<RegisterType, Long?>) {
+fun ExecuteLib.populateSnapShot(snapShotRegisters: MutableMap<RegisterType, Double?>) {
 	for (i in snapShotRegisters) {
 		registers.writeUnsafe(i.key, i.value)
 	}
