+++ src/main/kotlin/environment/libEx/snapShotRegisters.kt
@@ -4,8 +4,8 @@ import data.registers.RegisterType
 import environment.ExecuteLib
 import registers
 
-fun ExecuteLib.snapShotRegisters(): MutableMap<RegisterType, Long?> {
-	val allRegisters = mutableMapOf<RegisterType, Long?>()
+fun ExecuteLib.snapShotRegisters(): MutableMap<RegisterType, Double?> {
+	val allRegisters = mutableMapOf<RegisterType, Double?>()
 	for (i in RegisterType.entries) {
 		allRegisters[i] = registers.readUnsafe(i)
 	}
