+++ src/main/kotlin/internals/instructions/arithmetic/eq.kt
@@ -4,7 +4,7 @@ import data.registers.IntelRegisters
 import data.registers.RegisterType
 import data.registers.intelNames
 import errors
-import helpers.toLong
+import helpers.toDouble
 import registers
 
 /**
@@ -22,7 +22,7 @@ fun Arithmetic.eq(operand1: RegisterType, operand2: RegisterType) = try {
 	val out = registers.read(register = operand1) == registers.read(register = operand2)
 
 	registers.write(
-		intelNames[IntelRegisters.EF], if (out) true.toLong() else false.toLong()
+		intelNames[IntelRegisters.EF], if (out) true.toDouble() else false.toDouble()
 	)
 
 } catch (e: Exception) {
