+++ src/main/kotlin/internals/systemCalls/calls/exit.kt
@@ -14,6 +14,6 @@ import kotlin.system.exitProcess
  */
 @Suppress("RemoveExplicitTypeArguments")
 fun SystemCall.exit(s2: RegisterType) = call("exit") {
-    val exitCode: Long = registers.read(register = s2)
+    val exitCode: Long = registers.read(register = s2).toLong()
     exitProcess(status = with<Long, Int>(receiver = exitCode) { return@with this.run<Long, Int>(block = Long::toInt) })
 }
